sonar issue

+++ systemadmin/src/main/java/com/ecquaria/cloud/moh/iais/action/BlastManagementDelegator.java
@@ -400,7 +400,7 @@ public class BlastManagementDelegator {
             return errorMap;
         }
 
-        double size = (file.getSize() / 0x400) / (double) 0x400;
+        double size = (double)(file.getSize() / 0x400) / (double) 0x400;
 
         if (Math.ceil(size) > 0x10){
             errorMap.put(FILE_UPLOAD_ERROR, "GENERAL_ERR0004");
